pulling out feature flag utils
