+++ ui/apps/platform/src/Containers/Policies/Wizard/Step3/PolicyCriteriaForm.tsx
@@ -37,11 +37,7 @@ function PolicyCriteriaForm({ hasActiveViolations }: PolicyBehaviorFormProps) {
         }
     }
 
-    const filteredDescriptors = getPolicyDescriptors(
-        isFeatureFlagEnabled,
-        values.eventSource,
-        values.lifecycleStages
-    );
+    const filteredDescriptors = getPolicyDescriptors(values.eventSource, values.lifecycleStages);
 
     const headingElements = (
         <>
